requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-dotenv==1.0.0
motor==3.3.2
pymongo==4.6.1
pandas>=2.2.0  # Versión más reciente que soporta Python 3.13
openpyxl==3.1.2